Add custom cronet-common module to edit QUIC User-Agent ID
As Cronet doesn't allow ability to customize QUIC User-Agent ID, we need to
modify this behavior by using custom components.

This module is based on the official Chromium cronet-common module, with
modifications to change how the QUIC User-Agent ID is generated.

More precisely, the addition of the package name is removed and only the
Cronet part is kept.

Instructions to update this module have been added in the README of the module.

This custom module allows usage of QUIC in the app.